+++ src/theme/components/global.ts
@@ -36,6 +36,7 @@ const checkboxRadioControlSize = {
 const checkboxRadioHelperErrorTextStyle = {
   marginLeft: "30px", // calc(22px + var(--nypl-space-xs))
   _disabled: {
+    color: "ui.gray.dark",
     fontStyle: "italic",
   },
 };
@@ -62,6 +63,13 @@ const labelLegendText = {
     fontWeight: "helper.default",
   },
 };
+// Used for the disabled state of Select and TextInput
+const selectTextInputDisabledStyles = {
+  bg: "ui.disabled.secondary",
+  borderColor: "ui.disabled.primary",
+  color: "ui.gray.dark",
+  opacity: "1",
+};
 // Used for p, ul, and ol
 const textMargin = {
   margin: "0",
@@ -75,5 +83,6 @@ export {
   checkboxRadioHelperErrorTextStyle,
   checkboxRadioLabelStyles,
   labelLegendText,
+  selectTextInputDisabledStyles,
   textMargin,
 };