+++ src/theme/components/radio.ts
@@ -19,12 +19,10 @@ const baseStyleControl = {
     borderColor: "ui.focus",
     color: "ui.white",
     _disabled: {
-      color: "ui.gray.medium",
-      borderColor: "ui.gray.medium",
+      borderColor: "ui.disabled.primary",
       bg: "ui.white",
-
       _before: {
-        bg: "ui.gray.medium",
+        bg: "ui.disabled.primary",
       },
     },
     _invalid: {
@@ -49,7 +47,7 @@ const baseStyleControl = {
     },
   },
   _disabled: {
-    borderColor: "ui.gray.medium",
+    borderColor: "ui.disabled.primary",
     bg: "ui.white",
   },
   _focus: {